JS: remove useless externs definitions for tests
